+++ src/SMAPI/IAssetDataForMap.cs
@@ -17,5 +17,12 @@ namespace StardewModdingAPI
         /// <param name="targetArea">The tile area within the target map to overwrite, or <c>null</c> to patch the whole map. The original content within this area will be erased. This must be within the bounds of the existing map.</param>
         /// <param name="patchMode">Indicates how the map should be patched.</param>
         void PatchMap(Map source, Rectangle? sourceArea = null, Rectangle? targetArea = null, PatchMapMode patchMode = PatchMapMode.Overlay);
+
+        /// <summary>Extend the map if needed to fit the given size. Note that this is an expensive operation and resizes the map in-place.</summary>
+        /// <param name="map">The map to resize.</param>
+        /// <param name="minWidth">The minimum map width in tiles.</param>
+        /// <param name="minHeight">The minimum map height in tiles.</param>
+        /// <returns>Whether the map was resized.</returns>
+        bool ExtendMap(Map map, int minWidth, int minHeight);
     }
 }